feat: support long for N input

+++ Lab2/runtask.h
@@ -4,4 +4,4 @@
 
 #include "slice.h"
 
-int run_task(int n, IntSlice (*func)(int), char *label, int mpi_size);
+int run_task(long n, LongSlice (*func)(long), char *label, int mpi_size);
